--- src/components/ButtonArgs.ts
@@ -1,46 +1,0 @@
-import { StyleProps } from "./Button";
-
-export const mainButtonArgs: StyleProps = {
-  $backgroundColor: "#A860F6",
-  $hoverColor: "#D4B7F4",
-  $textColor: "white",
-  width: "12rem",
-  height: "3rem",
-  $textSize: "1rem",
-};
-
-export const authButtonArgs: StyleProps = {
-  $backgroundColor: "#D1D5DB",
-  $hoverColor: "#E5E7EB",
-  $textColor: "#374151",
-  width: "16rem",
-  height: "3rem",
-  $textSize: "1rem",
-};
-
-export const kakaoButtonArgs: StyleProps = {
-  $backgroundColor: "#facc15",
-  $hoverColor: "#fde047",
-  $textColor: "#374151",
-  width: "4rem",
-  height: "4rem",
-  $textSize: "1rem",
-};
-
-export const naverButtonArgs: StyleProps = {
-  $backgroundColor: "#10B981",
-  $hoverColor: "#34D399",
-  $textColor: "#374151",
-  width: "4rem",
-  height: "4rem",
-  $textSize: "1rem",
-};
-
-export const googleButtonArgs: StyleProps = {
-  $backgroundColor: "white",
-  $hoverColor: "#F4F5F7",
-  $textColor: "#374151",
-  width: "4rem",
-  height: "4rem",
-  $textSize: "1rem",
-};
