+++ src/styles/styled.ts
@@ -1,13 +1,13 @@
 import styled from "styled-components";
 
-export const Main = styled.main<{gap: string}>`
+export const Main = styled.main<{ gap?: string }>`
   display: flex;
   align-items: center;
   justify-content: center;
   min-height: 100vh;
   background-color: white;
   padding: 16px;
-  gap: ${(props) => props.gap};
+  gap: ${(props) => props.gap || "0px"};
   @media (max-width: 768px) {
     flex-direction: column;
   }
